Viterbi linear added.

+++ fb-pruner/src/algs_linear/fwdback_linear.h
@@ -10,7 +10,7 @@
 #ifndef _FWDBACK_LIN_H
 #define _FWDBACK_LIN_H
 
-float forward_Linear(const SEQUENCE*   query, 
+int forward_Linear(  const SEQUENCE*   query, 
                      const HMM_PROFILE* target, 
                      const int          Q, 
                      const int          T, 
@@ -18,7 +18,7 @@ float forward_Linear(const SEQUENCE*   query,
                      MATRIX_2D*         sp_MX,
                      float*             sc_final);
 
-float backward_Linear(const SEQUENCE*    query, 
+int backward_Linear( const SEQUENCE*    query, 
                      const HMM_PROFILE* target, 
                      const int          Q, 
                      const int          T,
