+++ fb-pruner/src/algs_quad/fwdback_quad.h
@@ -11,20 +11,20 @@
 #define _FWDBACK_QUAD_H
 
 
-float forward_Quad(const SEQUENCE*   query, 
+int forward_Quad( const SEQUENCE*   query, 
                   const HMM_PROFILE* target, 
                   const int          Q, 
                   const int          T, 
                   MATRIX_3D*         st_MX, 
                   MATRIX_2D*         sp_MX,
-                  float*             sc_final);
+                  float*             sc_final );
 
-float backward_Quad( const SEQUENCE*    query, 
+int backward_Quad(   const SEQUENCE*    query, 
                      const HMM_PROFILE* target, 
                      const int          Q, 
                      const int          T, 
                      MATRIX_3D*         st_MX, 
                      MATRIX_2D*         sp_MX,
-                     float*             sc_final);
+                     float*             sc_final );
 
 #endif /* _FWDBACK_QUAD_H */
